pyproject.toml
[project]
name = "daily-planner-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-oauthlib>=1.2.2",
    "google-generativeai>=0.8.6",
    "langchain-google-genai>=4.1.3",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
    "todoist-api-python>=3.1.0",
]
